optimize by caching len lookups of lists within the large list
Instead of having to do an O(n) lookup in the list for the item in each
iteration, first cache the length of all sections in a dict, providing
O(1) lookups. This way, only the necessary elements need the O(n) list
lookup. For the current dataset, improved performance by ~60%.